server/analytics-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic==2.6.3
PyJWT>=2.8